% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geologic_data.R
\name{add_extra_geologic_columns}
\alias{add_extra_geologic_columns}
\title{Add extra Geologic Frac Database columns at a well level}
\usage{
add_extra_geologic_columns(well_comp_df, frac_df)
}
\arguments{
\item{well_comp_df}{well-level dataframe from `summarize_geologic_frac_data()`}

\item{frac_df}{stage-level dataframe from `load_process_geologic_frac_data()`}
}
\value{
dataframe of well level statistics with extra columns
}
\description{
This function adds columns that aren't available in the BCOGC database
to the well dataframe generated by `summarize_geologic_frac_data()`
}
