% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{b_value_plot}
\alias{b_value_plot}
\title{Create a plot from the b-value analysis}
\usage{
b_value_plot(ecdf, cat, out_df)
}
\arguments{
\item{ecdf}{dataframe with ecdf}

\item{cat}{dataframe with earthquake catalogue}

\item{out_df}{dataframe with model results}
}
\value{
dataframe with model output
}
\description{
Create a plot from the b-value analysis
}
