% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{calc_ecdf}
\alias{calc_ecdf}
\title{ECDF Calculation using Freedman & Diaconis (1981) approach for probability 
mass function for an earthquake catalogue.}
\usage{
calc_ecdf(cat, mwcol = "mw", k_bins = 10)
}
\arguments{
\item{cat}{earthquake catalogue}

\item{mwcol}{string denoting catalogue magnitude column}
}
\description{
ECDF Calculation using Freedman & Diaconis (1981) approach for probability 
mass function for an earthquake catalogue.
}
