% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{calc_skipped_stages}
\alias{calc_skipped_stages}
\title{Estimate skipped stages from BCOGC database}
\usage{
calc_skipped_stages(frac_df)
}
\arguments{
\item{frac_df}{processed dataframe from load_process_bcogc_frac_data()}
}
\value{
column of skipped stages for input dataframe
}
\description{
Calculates skipped stages based on low volume injections. Not perfect 
as need the 'planned' stages for each well which aren't available in the BCOGC 
catalogue (vs. the Geologic Frac Database)
}
