% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_analysis.R
\name{calculate_horizontal_well_spacing}
\alias{calculate_horizontal_well_spacing}
\title{Calculate horizontal well spacing and join with horizontal well dataframe}
\usage{
calculate_horizontal_well_spacing(horiz_dev_well_sf)
}
\arguments{
\item{horiz_dev_well_sf}{dataframe of montney wells}
}
\value{
a joined sf dataframe with distances added
}
\description{
Calculate horizontal well spacing and join with horizontal well dataframe
}
