% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{cat_bvalue_plot}
\alias{cat_bvalue_plot}
\title{Evaluate b-value models and make a plot}
\usage{
cat_bvalue_plot(cat, label)
}
\arguments{
\item{cat}{earthquake catalogue}

\item{label}{string for labelling plot}
}
\value{
ggplot
}
\description{
Evaluate b-value models and make a plot
}
