% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_analysis.R
\name{filter_extents_add_on_prod_comp}
\alias{filter_extents_add_on_prod_comp}
\title{Filter extents of wells and add in production and well completions}
\usage{
filter_extents_add_on_prod_comp(sf_df, montney_extents, well_prod_df,
  well_comp_df)
}
\arguments{
\item{sf_df}{an sf dataframe to be filtered and modified}

\item{montney_extents}{an sf polygon that will contain the study data}

\item{well_prod_df}{production data}

\item{well_comp_df}{completion data}
}
\value{
a filtered and modified dataframe
}
\description{
Filter extents of wells and add in production and well completions
}
