% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{get_deviated_surveys}
\alias{get_deviated_surveys}
\title{Get surveys for all wells that aren't 'horizontal'}
\usage{
get_deviated_surveys(well_survey_sf, horizontal_well_sf)
}
\arguments{
\item{well_survey_sf}{survey dataframe}

\item{horizontal_well_sf}{horizontal survey dataframe from `get_horizontal_laterals`}
}
\value{
horizontal laterals as sf linestrings
}
\description{
Get surveys for all wells that aren't 'horizontal'
}
