% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismogenic_association.R
\name{get_earthquake_stats}
\alias{get_earthquake_stats}
\title{Function to map the max, mean earthquake magnitudes and count the number of earthquakes}
\usage{
get_earthquake_stats(well_sf, is_catalog)
}
\arguments{
\item{well_sf}{sf dataframe for the well}

\item{is_catalog}{the filtered seismic catalog}
}
\description{
Function to map the max, mean earthquake magnitudes and count the number of earthquakes
}
