% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{get_horiz_dev_sf}
\alias{get_horiz_dev_sf}
\title{Wrapper to get horizontal and deviated wells from the well survey sf}
\usage{
get_horiz_dev_sf(horiz_dev_sf)
}
\arguments{
\item{horiz_dev_sf}{processed surveys from `load_process_ogc_survey_data`}
}
\value{
sf dataframe of horizontal and deviated wells
}
\description{
Wrapper to get horizontal and deviated wells from the well survey sf
}
