% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{get_horizontal_laterals}
\alias{get_horizontal_laterals}
\title{Calculate the lateral portion of horizontals, using an aggressive inclination cutoff 
to avoid deviated wells and get a good mean tvd}
\usage{
get_horizontal_laterals(survey_df)
}
\arguments{
\item{survey_df}{survey dataframe}
}
\value{
horizontal laterals as sf linestrings
}
\description{
Calculate the lateral portion of horizontals, using an aggressive inclination cutoff 
to avoid deviated wells and get a good mean tvd
}
