% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismogenic_association.R
\name{get_mean_mag}
\alias{get_mean_mag}
\title{Mapped function to get mean magnitude}
\usage{
get_mean_mag(earthquake, is_catalog)
}
\arguments{
\item{earthquake}{the earthquake list in question}

\item{is_catalog}{the filtered seismic catalog}
}
\description{
Mapped function to get mean magnitude
}
