% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismogenic_association.R
\name{get_n_earthquakes}
\alias{get_n_earthquakes}
\title{Mapped function to count number of earthquakes for each well}
\usage{
get_n_earthquakes(earthquake, is_catalog)
}
\arguments{
\item{earthquake}{the earthquake list in question}

\item{is_catalog}{the filtered seismic catalog}
}
\description{
Mapped function to count number of earthquakes for each well
}
