% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{gr_prob}
\alias{gr_prob}
\title{Gutenberg Richter Probability}
\usage{
gr_prob(model, M = 2)
}
\arguments{
\item{model}{dataframe with b value}
}
\value{
vector of probabilities
}
\description{
Returns probability of earthquakes above magnitude zero. Can
accept a single value of M or a vector.
}
