% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismic_catalog.R
\name{hdbscan_seismic_catalog}
\alias{hdbscan_seismic_catalog}
\title{Cluster earthquake catalog using hdbscan}
\usage{
hdbscan_seismic_catalog(seismic_catalog, npts = 5)
}
\arguments{
\item{seismic_catalog}{a mw_vector of magnitudes}

\item{npts}{The minimum number of points to start a cluster}
}
\value{
a dataframe with catalog centres and an ecdf, filtered to above the magnitude of completness
}
\description{
Cluster earthquake catalog using hdbscan
}
