% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seismic_catalog.R, R/seismic_catalog.R
\name{k_bins}
\alias{k_bins}
\title{Function to determine bin numbers for catalog determination}
\usage{
k_bins(mw_vector, method = "rice")

k_bins(mw_vector, method = "rice")
}
\arguments{
\item{mw_vector}{a mw_vector of magnitudes}

\item{method}{a character of methods: 'rice', 'sturges', 'freedman'}

\item{mw_vector}{a mw_vector of magnitudes}

\item{method}{a character of methods: 'rice', 'sturges', 'freedman'}
}
\value{
an integer number of bins

an integer number of bins
}
\description{
The ECDF function requires binning earthquakes into discrete categories. This
complicates our modelling process by turning a continuous distribution into a 
discrete categorical series of bins. There are There are several methods for determining
this binning: 
Sturges, H. (1926) The choice of a class-interval. J. Amer. Statist. Assoc., 21, 65-66.
Scott, D.W. (1979) On optimal and data-based histograms. Biometrika, 66, 605-610.
Freedman, D. and Diaconis, P. (1981) On this histogram as a density estimator: L2 theory. Zeit. Wahr. ver. Geb., 57, 453-476.
Wand, M. P. (1997). Data-based choice of histogram bin width. The American Statistician, 51(1), 59-64.
We use three common methods to calculate the deterministic b-value of 
the catalogue. These functions are also used in the bootstrap estimation
of uncertainity.

The ECDF function requires binning earthquakes into discrete categories. This
complicates our modelling process by turning a continuous distribution into a 
discrete categorical series of bins. There are There are several methods for determining
this binning: 
Sturges, H. (1926) The choice of a class-interval. J. Amer. Statist. Assoc., 21, 65-66.
Scott, D.W. (1979) On optimal and data-based histograms. Biometrika, 66, 605-610.
Freedman, D. and Diaconis, P. (1981) On this histogram as a density estimator: L2 theory. Zeit. Wahr. ver. Geb., 57, 453-476.
Wand, M. P. (1997). Data-based choice of histogram bin width. The American Statistician, 51(1), 59-64.
We use three common methods to calculate the deterministic b-value of 
the catalogue. These functions are also used in the bootstrap estimation
of uncertainity.
}
