% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismic_catalog.R
\name{load_preprocessed_catalogue}
\alias{load_preprocessed_catalogue}
\title{Load preprocessed seismic catalogue}
\usage{
load_preprocessed_catalogue(file)
}
\arguments{
\item{file}{string of file path}
}
\value{
an sf dataframe
}
\description{
Load preprocessed seismic catalogue
}
