% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geologic_data.R
\name{load_process_geologic_frac_data}
\alias{load_process_geologic_frac_data}
\title{Load completions data from Geologic Frac Database}
\usage{
load_process_geologic_frac_data(file)
}
\arguments{
\item{file}{input file, downloaded from Geologic Frac Database (proprietary)}
}
\value{
dataframe
}
\description{
loads from a file using fread, followed by some basic filtering and cleaning,
and then a lot of case_when to clean the input data to a consistent format for future
algorithms
}
