% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{load_process_ogc_prod_data}
\alias{load_process_ogc_prod_data}
\title{Load and process production data}
\usage{
load_process_ogc_prod_data(file)
}
\arguments{
\item{file}{file for production data from BC OGC}
}
\value{
granular (month by month) production data
}
\description{
Load and process production data
}
