% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{load_process_ogc_surface_data}
\alias{load_process_ogc_surface_data}
\title{Load and process BCOGC surface survey data}
\usage{
load_process_ogc_surface_data(file)
}
\arguments{
\item{file}{file for survey data from BC OGC}
}
\value{
sf dataframe of well surface information
}
\description{
Load and process BCOGC surface survey data
}
