% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{load_process_ogc_survey_data}
\alias{load_process_ogc_survey_data}
\title{Load and process BCOGC downhole survey data for deviated and horizontal wells}
\usage{
load_process_ogc_survey_data(file, well_surface_sf)
}
\arguments{
\item{file}{file for deviated / horizontal survey data from BC OGC}

\item{well_surface_sf}{surface survey information from BCOGC from `load_process_ogc_surface_data()`}
}
\value{
sf dataframe of well surveys with lines of deviated/horizontal wells
}
\description{
Load and process BCOGC downhole survey data for deviated and horizontal wells
}
