% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{load_process_vertical_surveys}
\alias{load_process_vertical_surveys}
\title{Get single points for the remaining wells based on surface location}
\usage{
load_process_vertical_surveys(file, well_surface_sf, horiz_dev_wa_num)
}
\arguments{
\item{file}{filename}

\item{well_surface_sf}{surface survey sf dataframe}

\item{horiz_dev_wa_num}{numeric vector of wa_num from horizontal wells}
}
\value{
horizontal laterals as sf linestrings
}
\description{
Get single points for the remaining wells based on surface location
}
