% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_analysis.R
\name{load_resample_geotiffs}
\alias{load_resample_geotiffs}
\title{Load and process a folder of geotiffs}
\usage{
load_resample_geotiffs(file_path, extent_sf, res = c(500, 500))
}
\arguments{
\item{file_path}{full path to geotiffs}

\item{extent_sf}{the polygon used to mask and resample the rasters (the study extent)}

\item{res}{the desired resolution for resampling, 500m x 500m by default}
}
\value{
a raster stack for property extract
}
\description{
Load and process a folder of geotiffs
}
