% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{make_pdp_plot}
\alias{make_pdp_plot}
\title{Make a single PDP plot for mapping}
\usage{
make_pdp_plot(i, predictor, feats)
}
\arguments{
\item{i}{feature number}

\item{predictor}{predictor}

\item{feats}{feats}
}
\value{
ggplot
}
\description{
Make a single PDP plot for mapping
}
