% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_analysis.R
\name{mapped_horizontal_well_spacing}
\alias{mapped_horizontal_well_spacing}
\title{Function to be mapped and get the horizontal well spacing and wells within a certain
distance}
\usage{
mapped_horizontal_well_spacing(i, horiz_montney_wells, nn_dist)
}
\arguments{
\item{i}{dataframe row iter}

\item{horiz_montney_wells}{dataframe of montney wells}

\item{nn_dist}{nearest neighbour distances (pre-calculated)}
}
\value{
a row of a dataframe, to be reduced into a proper df
}
\description{
Function to be mapped and get the horizontal well spacing and wells within a certain
distance
}
