% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ml_corrplot}
\alias{ml_corrplot}
\title{Make a correlation plot}
\usage{
ml_corrplot(ml_df, output_path = "../output/", features, target)
}
\arguments{
\item{output_path}{output path for plot}

\item{num_ml_df}{clean, numeric only dataframe of ml predictors}
}
\value{
ggplot
}
\description{
Make a correlation plot
}
