% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seismic_catalog.R, R/seismic_catalog.R
\name{mmaxc}
\alias{mmaxc}
\title{Function to determine the maximum magnitude of a catalog}
\usage{
mmaxc(centres, rev_ecdf)

mmaxc(centres, rev_ecdf)
}
\arguments{
\item{centres}{bin centres}

\item{rev_ecdf}{reverse ecdf used for gutenberg richter relationship}

\item{centres}{bin centres}

\item{rev_ecdf}{reverse ecdf used for gutenberg richter relationship}
}
\value{
double maximum magnitude of distribution

double maximum magnitude of distribution
}
\description{
Function to determine the maximum magnitude of a catalog

Function to determine the maximum magnitude of a catalog
}
