% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{most_freq_var}
\alias{most_freq_var}
\title{Return most frequent variable from a dataframe}
\usage{
most_freq_var(df, var)
}
\arguments{
\item{df}{dataframe to interogate}

\item{var}{string of variable name to roll up}
}
\value{
column from summarized dataframe (using pull)
}
\description{
Meant to be used in a dplyr::mutate call to get most frequent occurence 
of a variable by wa_number
}
