% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pca_scree_plot}
\alias{pca_scree_plot}
\title{Calculate and Plot a PCA scree plot}
\usage{
pca_scree_plot(ml_df, output_path)
}
\arguments{
\item{output_path}{output path for plot}

\item{num_ml_df}{clean, numeric only dataframe of ml predictors}
}
\value{
ggplot
}
\description{
Calculate and Plot a PCA scree plot
}
