% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_classif_feat_rand}
\alias{plot_classif_feat_rand}
\title{Plot wrapper feature selection results}
\usage{
plot_classif_feat_rand(path, output_path = "../output/",
  n_best_val = 10)
}
\arguments{
\item{output_path}{output path for plot}

\item{n_best_val}{number of top values to compute}

\item{num_ml_df}{clean, numeric only dataframe of ml predictors}
}
\value{
ggplot
}
\description{
Plot wrapper feature selection results
}
