% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{GGPlot function to plot an ECDF with minimal formatting}
\usage{
plot_ecdf(ecdf, out_file = "ecdf")
}
\arguments{
\item{ecdf}{ecdf dataframe with centres, cdf, and log_cdf}
}
\description{
GGPlot function to plot an ECDF with minimal formatting
}
