% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_regr_feat_seq}
\alias{plot_regr_feat_seq}
\title{Plot wrapper feature selection results}
\usage{
plot_regr_feat_seq(path, output_path = "../output/", n_best_val = 1000)
}
\arguments{
\item{output_path}{output path for plot}

\item{num_ml_df}{clean, numeric only dataframe of ml predictors}
}
\value{
ggplot
}
\description{
Plot wrapper feature selection results
}
