% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_regr_filt_importance}
\alias{plot_regr_filt_importance}
\title{Make a target scatter plot}
\usage{
plot_regr_filt_importance(regr_tsk, output_path = "../output/")
}
\arguments{
\item{output_path}{output path for plot}

\item{num_ml_df}{clean, numeric only dataframe of ml predictors}
}
\value{
ggplot
}
\description{
Make a target scatter plot
}
