% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seismic_catalog.R, R/seismic_catalog.R
\name{process_cat}
\alias{process_cat}
\title{Function to process an earthquake catalog, as a vector of magnitudes}
\usage{
process_cat(mw_vector, method = "rice")

process_cat(mw_vector, method = "rice")
}
\arguments{
\item{mw_vector}{a mw_vector of magnitudes}

\item{method}{a character of methods: 'rice', 'sturges', 'freedman'}

\item{mw_vector}{a mw_vector of magnitudes}

\item{method}{a character of methods: 'rice', 'sturges', 'freedman'}
}
\value{
a dataframe with catalog centres and an ecdf, filtered to above the magnitude of completness

a dataframe with catalog centres and an ecdf, filtered to above the magnitude of completness
}
\description{
Function to process an earthquake catalog, as a vector of magnitudes

Function to process an earthquake catalog, as a vector of magnitudes
}
