% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{run_plot_models}
\alias{run_plot_models}
\title{Run three models to fit distribution, output parameters, and plot results}
\usage{
run_plot_models(ecdf, cat, plot = FALSE)
}
\arguments{
\item{ecdf}{dataframe with ecdf}

\item{cat}{dataframe with earthquake catalogue}
}
\value{
dataframe with model output
}
\description{
Run three models to fit distribution, output parameters, and plot results
}
