% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seismogenic_association.R
\name{run_spatiotemporal_filter}
\alias{run_spatiotemporal_filter}
\title{Spatiotemporal Association Filter}
\usage{
run_spatiotemporal_filter(well_sf, seismic_catalog, dist = 5000,
  days_before = 1, days_after = 30, mw_min = 1.1)
}
\arguments{
\item{well_sf}{the sf dataframe of the wells}

\item{seismic_catalog}{the filtered seismic catalog}

\item{dist}{distance to well}

\item{days_before}{days before start of frac}

\item{days_after}{days after the end of the frac}

\item{mw_min}{minimum magnitude}
}
\value{
the input dataframe (well_sf) with seismogenic indicators
}
\description{
Run the Schultz et al. (2017) spatiotemporal filter to flag wells & 
associated earthquakes that are a) within a cluster, b) within x days of injection 
starting, c) prior to x days after injection stops, d) within x m, and e) larger than 
a minimum magnitude (generally the magnitude of completeness)
}
