% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{st_line_midpoints}
\alias{st_line_midpoints}
\title{Calculate the midpoint of lines for horizontal laterals, for mapping}
\usage{
st_line_midpoints(sf_lines = NULL)
}
\arguments{
\item{sf_lines}{sf linestring}
}
\value{
midpoint as sf point
}
\description{
Calculate the midpoint of lines for horizontal laterals, for mapping
}
