% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{summarize_bcogc_frac_data}
\alias{summarize_bcogc_frac_data}
\title{Summarize BCOGC Frac Database at a well level}
\usage{
summarize_bcogc_frac_data(frac_df)
}
\arguments{
\item{frac_df}{processed dataframe from load_process_bcogc_frac_data()}
}
\value{
dataframe of well level statistics
}
\description{
This function summarizes the Frac Database at a well level, in a consistent
way relative to the geoLOGIC data. Has lots of warnings due to missing values, but can't
explicitly remove errors without killing a lot of data
}
