% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcogc_data.R
\name{summarize_prod_data}
\alias{summarize_prod_data}
\title{Summarize production data by well}
\usage{
summarize_prod_data(prod_df)
}
\arguments{
\item{prod_df}{dataframe of production data, from `load_process_ogc_prod_data()`}
}
\value{
dataframe of well by well production data rollup
}
\description{
Summarize production data by well
}
