% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{target_boxplot}
\alias{target_boxplot}
\title{Make a boxplot of variables against the target}
\usage{
target_boxplot(ml_df, output_path = "../output/", prefix = "")
}
\arguments{
\item{ml_df}{clean, numeric only dataframe of ml predictors}

\item{output_path}{output path for plot}
}
\value{
ggplot
}
\description{
Make a boxplot of variables against the target
}
