% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{target_scatter_plot}
\alias{target_scatter_plot}
\title{Make a target scatter plot}
\usage{
target_scatter_plot(ml_df, output_path = "../output/")
}
\arguments{
\item{output_path}{output path for plot}

\item{num_ml_df}{clean, numeric only dataframe of ml predictors}
}
\value{
ggplot
}
\description{
Make a target scatter plot
}
