% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{train_aki_model}
\alias{train_aki_model}
\title{Aki (1965) method}
\usage{
train_aki_model(ecdf, cat)
}
\arguments{
\item{ecdf}{dataframe with ecdf}
}
\value{
dataframe with model output
}
\description{
Function to calculate the maximum curvature Mc, 'a' value (intercept),
and b-value using the Aki (1965) method.
}
