% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{train_gof_model}
\alias{train_gof_model}
\title{Goodness of Fit Method}
\usage{
train_gof_model(ecdf)
}
\arguments{
\item{ecdf}{dataframe with ecdf}
}
\value{
dataframe with model output
}
\description{
Optimization algorithm to fit three dimensional goodness of 
fit b-value.
}
