% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_catalog_plot_functions.R
\name{train_maxc_model}
\alias{train_maxc_model}
\title{Maximum curvature method linear model}
\usage{
train_maxc_model(ecdf)
}
\arguments{
\item{ecdf}{dataframe with ecdf}
}
\value{
dataframe with model output
}
\description{
Function to calculate the maximum curvature Mc, a value (intercept),
and b-value using a smoothing spline and predictor method.
}
